[Front] migration to tsx of : Indicator, IndicatorsLines (#2952)
